[feat] 更新 JokerOption 的历史时间范围，重构服务以使用 IDbContextFactory，优化数据准备流程

+++ Services/InterestService.cs
@@ -48,7 +48,7 @@ internal class InterestService(IBybitRestClient restClient, JokerContext context
 
         var interests = interestResult.Data.List.Select(r => new T {
             Timestamp = r.Timestamp,
-            OpenInterest = r.OpenInterest,
+            OpenInterest = (double)r.OpenInterest,
             Symbol = symbol
         }).ToArray();
